refactor(tui::ui::components::popups): move "MessagePopup" to own file
