+++ tui/src/ui/components/popups/mod.rs
@@ -25,17 +25,16 @@
  */
 use crate::ui::Msg;
 use termusicplayback::SharedSettings;
-use tui_realm_stdlib::{Paragraph, Radio};
+use tui_realm_stdlib::Radio;
 use tuirealm::command::{Cmd, CmdResult, Direction};
 use tuirealm::event::{Key, KeyEvent, NoUserEvent};
-use tuirealm::props::{
-    Alignment, BorderType, Borders, Color, PropPayload, PropValue, TextModifiers, TextSpan,
-};
+use tuirealm::props::{Alignment, BorderType, Borders, Color, PropPayload, PropValue};
 use tuirealm::{AttrValue, Attribute, Component, Event, MockComponent, State, StateValue};
 
 mod deleteconfirm;
 mod error;
 mod help;
+mod message;
 mod podcast;
 mod saveplaylist;
 
@@ -43,6 +42,7 @@ mod saveplaylist;
 pub use deleteconfirm::{DeleteConfirmInputPopup, DeleteConfirmRadioPopup};
 pub use error::ErrorPopup;
 pub use help::HelpPopup;
+pub use message::MessagePopup;
 #[allow(unused_imports)]
 pub use podcast::{
     FeedDeleteConfirmInputPopup, FeedDeleteConfirmRadioPopup, PodcastAddPopup,
@@ -164,38 +164,5 @@ impl Component<Msg, NoUserEvent> for QuitPopup {
         } else {
             Some(Msg::None)
         }
-    }
-}
-
-#[derive(MockComponent)]
-pub struct MessagePopup {
-    component: Paragraph,
-}
-
-impl MessagePopup {
-    pub fn new<S: AsRef<str>>(title: S, msg: S) -> Self {
-        Self {
-            component: Paragraph::default()
-                .borders(
-                    Borders::default()
-                        .color(Color::Cyan)
-                        .modifiers(BorderType::Rounded),
-                )
-                .foreground(Color::Green)
-                // .background(Color::Black)
-                .modifiers(TextModifiers::BOLD)
-                .alignment(Alignment::Center)
-                .title(title, Alignment::Center)
-                .text(vec![TextSpan::from(msg.as_ref().to_string())].as_slice()),
-        }
-    }
-}
-
-impl Component<Msg, NoUserEvent> for MessagePopup {
-    fn on(&mut self, _ev: Event<NoUserEvent>) -> Option<Msg> {
-        // match ev {
-        //     _ => None,
-        // }
-        None
     }
 }
